Add all API methods in the service. Created pharmacies map activity. It can get user's location and query nearby pharmacies.

+++ Calendula/src/main/java/es/usc/citius/servando/calendula/pharmacies/remote/PharmaciesService.java
@@ -74,5 +74,5 @@ public interface PharmaciesService {
      * @return
      */
     @GET("location/{lat}/{long}/{distance}/{open}")
-    Call<List<Pharmacy>> listByLocation(@Path("lat") Float latitude, @Path("long") Float longitude, @Path("distance") Integer distance, @Path("open") String open);
+    Call<List<Pharmacy>> listByLocation(@Path("lat") Double latitude, @Path("long") Double longitude, @Path("distance") Integer distance, @Path("open") String open);
 }
